+++ src/main/java/com/codeblooded/travelbookingsystem/service/EmailService.java
@@ -60,7 +60,7 @@ public class EmailService {
         }
     }
 
-    public void sendBookingConfirmationEmail(String emailAddress, int customerID, int bookingID, int paymentID, int travelPackageID, String departureDate, String bookingStatus) {
+    public void sendBookingConfirmationEmail(String emailAddress, long customerID, int bookingID, int paymentID, int travelPackageID, String departureDate, String bookingStatus) {
         String templateFilePath = "Email-Templates/BookingConfirmationTemplate.html";
         String subject = "Concordia Travel Booking System: Booking Confirmation";
         BookingConfirmationTemplateData bookingConfirmationTemplateData = new BookingConfirmationTemplateData(customerID, bookingID, paymentID, travelPackageID, departureDate, bookingStatus);
@@ -78,7 +78,7 @@ public class EmailService {
         }
     }
 
-    public void sendBookingUpdateEmail(String emailAddress, int customerID, int bookingID, int travelPackageID, String departureDate, String bookingStatus) {
+    public void sendBookingUpdateEmail(String emailAddress, long customerID, int bookingID, int travelPackageID, String departureDate, String bookingStatus) {
         String templateFilePath = "Email-Templates/BookingUpdateTemplate.html";
         String subject = "Concordia Travel Booking System: Booking Confirmation";
         BookingUpdateTemplateData bookingUpdateTemplateData = new BookingUpdateTemplateData(customerID, bookingID, travelPackageID, departureDate, bookingStatus);
@@ -110,7 +110,7 @@ public class EmailService {
     }
 
     private static class BookingConfirmationTemplateData {
-        private final int customerID;
+        private final long customerID;
         private final int bookingID;
         private final int paymentID;
         private final int travelPackageID;
@@ -118,7 +118,7 @@ public class EmailService {
         private final String bookingStatus;
 
 
-        private BookingConfirmationTemplateData(int customerID, int bookingID, int paymentID, int travelPackageID, String departureDate, String bookingStatus) {
+        private BookingConfirmationTemplateData(long customerID, int bookingID, int paymentID, int travelPackageID, String departureDate, String bookingStatus) {
             this.customerID = customerID;
             this.bookingID = bookingID;
             this.paymentID = paymentID;
@@ -127,7 +127,7 @@ public class EmailService {
             this.bookingStatus = bookingStatus;
         }
 
-        public int getCustomerID() {
+        public long getCustomerID() {
             return customerID;
         }
 
@@ -153,14 +153,14 @@ public class EmailService {
     }
 
     private static class BookingUpdateTemplateData {
-        private final int customerID;
+        private final long customerID;
         private final int bookingID;
         private final int travelPackageID;
         private final String departureDate;
 
         private final String bookingStatus;
 
-        private BookingUpdateTemplateData(int customerID, int bookingID, int travelPackageID, String departureDate, String bookingStatus) {
+        private BookingUpdateTemplateData(long customerID, int bookingID, int travelPackageID, String departureDate, String bookingStatus) {
             this.customerID = customerID;
             this.bookingID = bookingID;
             this.travelPackageID = travelPackageID;
@@ -168,7 +168,7 @@ public class EmailService {
             this.bookingStatus = bookingStatus;
         }
 
-        public int getCustomerID() {
+        public long getCustomerID() {
             return customerID;
         }
